--- src/models.js
@@ -1,35 +1,0 @@
-export class Profile {
-    constructor(
-        slug,
-        name,
-        avatarUrl
-    ) { }
-}
-
-export class Company {
-    constructor(
-        slug,
-        thread
-    ) { }
-    // createdBy: {
-    //     slug: "",
-    //     name: "Dev User",
-    //     avatarUrl: null
-    // },
-    // thread: Number
-    // createdAt: "2020-08-25T07:45:25.992590Z",
-    // clapCount: 0,
-    // threadSize: null,
-    // name: "Hypar",
-    // description: "We provide a cloud compute environment and software libraries that enable AEC developers to deploy their specialized building logic to the cloud, and make it available to their team, their organization, or the world.",
-    // website: "https://hypar.io/",
-    // location: "Culver City, CA, USA",
-    // twitterHandle: null,
-    // crunchbaseId: null,
-    // logoUrl: "https://aecworks-bucket-local.s3.amazonaws.com/images/a3853cb6f9c54bad8e0dea700f178e22.png",
-    // coverUrl: "https://aecworks-bucket-local.s3.amazonaws.com/images/a3853cb6f9c54bad8e0dea700f178e22.png",
-    // hashtags: [
-    //     "webapp",
-    //     "generative"
-    // ]
-}
